Refacto: add GenderEditor
