+++ src/views/components/database/pokemon/editors/EncounterEditor.tsx
@@ -2,7 +2,7 @@ import { Editor } from '@components/editor/Editor';
 import { EditorHandlingClose, useEditorHandlingClose } from '@components/editor/useHandleCloseEditor';
 import { useCreaturePage } from '@hooks/usePage';
 import { useSelectOptions } from '@hooks/useSelectOptions';
-import React, { forwardRef, useMemo, useRef } from 'react';
+import React, { forwardRef, useMemo } from 'react';
 import { useTranslation } from 'react-i18next';
 import { useUpdateForm } from './useUpdateForm';
 import { useZodForm } from '@hooks/useZodForm';
@@ -11,6 +11,7 @@ import { InputFormContainer } from '@components/inputs/InputContainer';
 import { ENCOUNTER_EDITOR_SCHEMA } from './EncounterEditor/EncounterEditorSchema';
 import { ItemHeldEditor } from './EncounterEditor/ItemHeldEditor';
 import { StudioCreatureForm } from '@modelEntities/creature';
+import { GenderEditor } from './EncounterEditor/GenderEditor';
 
 const initForm = (form: StudioCreatureForm) => {
   return {
@@ -27,9 +28,8 @@ export const EncounterEditor = forwardRef<EditorHandlingClose>((_, ref) => {
   const updateForm = useUpdateForm(creature, form);
   const itemOptions = useSelectOptions('itemHeld');
   const options = useMemo(() => [{ value: 'none', label: tSelect('none') }, ...itemOptions], []);
-  const divRef = useRef<HTMLDivElement>(null);
   const { canClose, getFormData, getRawFormData, onInputTouched: onTouched, defaults, formRef } = useZodForm(ENCOUNTER_EDITOR_SCHEMA, initForm(form));
-  const { Input, EmbeddedUnitInput, Toggle } = useInputAttrsWithLabel(ENCOUNTER_EDITOR_SCHEMA, defaults);
+  const { Input } = useInputAttrsWithLabel(ENCOUNTER_EDITOR_SCHEMA, defaults);
 
   const onClose = () => {
     const result = canClose() && getFormData();
@@ -42,25 +42,11 @@ export const EncounterEditor = forwardRef<EditorHandlingClose>((_, ref) => {
   };
   useEditorHandlingClose(ref, onClose, canClose);
 
-  const isGenderless = Boolean(getRawFormData().isGenderLess ?? defaults.isGenderLess === 'true');
-
-  const onStateChange: React.ChangeEventHandler<HTMLInputElement> = ({ currentTarget }) => {
-    const femaleRateElement = formRef.current?.elements.namedItem('femaleRate');
-    if (!(femaleRateElement instanceof HTMLInputElement)) return;
-
-    femaleRateElement.value = '0';
-    const isGenderless = currentTarget.checked;
-    if (divRef.current) divRef.current.style.display = isGenderless ? 'none' : 'block';
-  };
-
   return (
     <Editor type="edit" title={t('encounter')}>
       <InputFormContainer ref={formRef}>
         <Input name="catchRate" label={t('catch_rate')} labelLeft onInput={onTouched} />
-        <Toggle name="isGenderLess" label={t('genderless')} onChange={onStateChange} onInput={onTouched} />
-        <div style={{ display: isGenderless ? 'none' : undefined }} ref={divRef}>
-          <EmbeddedUnitInput name="femaleRate" unit="%" type="number" label={t('female_rate')} labelLeft onInput={onTouched} />
-        </div>
+        <GenderEditor getRawFormData={getRawFormData} defaults={defaults} onTouched={onTouched} formRef={formRef} />
         <ItemHeldEditor index={0} options={options} getRawFormData={getRawFormData} defaults={defaults} onTouched={onTouched} />
         <ItemHeldEditor index={1} options={options} getRawFormData={getRawFormData} defaults={defaults} onTouched={onTouched} />
       </InputFormContainer>
